fix(): move build logic to parameters-container
